pyproject.toml
[project]
name = "podcast-ad-cutter"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "feedparser>=6.0",
    "httpx>=0.27",
    "litellm>=1.40",
    "openai>=1.30",
    "pydub>=0.25",
    "ffmpeg-python>=0.2",
    "aiosqlite>=0.20",
    "pyyaml>=6.0",
    "pydantic>=2.7",
    "tenacity>=8.3",
    "rich>=13.7",
    "python-dotenv>=1.0",
    "fastapi>=0.133.1",
    "uvicorn[standard]>=0.41.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.22",
    "sse-starlette>=3.2.0",
    "watchfiles>=1.1.1",
    "anyio>=4.12.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "models*", "pipeline*", "db*", "frontend*"]
exclude = ["tests*", "output*", "data*"]

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "mypy>=1.10",
    "ruff>=0.4",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "ALL",
]
ignore = [
    "D100",     # missing module docstring
    "D104",     # missing package docstring
    "D107",     # missing __init__ docstring
    "D203",     # conflicts with D211 (no-blank-line-before-class wins)
    "D213",     # conflicts with D212 (multi-line-summary-first-line wins)
    "TRY003",   # exception message in raise
    "EM102",    # f-string in exception
    "COM812",   # trailing comma
    "G004",     # f-string in logging
    "PLR0913",  # too many function arguments
    "PLW0603",  # global statement (module singleton pattern)
    "BLE001",   # broad exception catch (intentional in error handlers)
    "FBT001",   # boolean-typed positional argument
    "FBT002",   # boolean default positional argument
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "ANN",
    "S101",
    "PLC0415",
    "D103",
    "SLF001",
    "PLR2004",
    "INP001",
    "ASYNC240", # test fixtures use sync path reads; anyio.Path not needed in tests
]
"frontend/app.py" = ["PLC0415"]
"frontend/routes/pipeline.py" = ["PLC0415"]

[tool.mypy]
python_version = "3.12"
strict = true
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
